requirements.txt
Flask==3.1.0
matplotlib==3.9.3
networkx==3.4.2
numpy==2.0.2
pandas==2.2.3
Pillow==11.0.0
plotly==6.0.0
pyspark==3.5.3
wordcloud==1.9.4
